Bigger money storage (#174)
* сделал так, что переменные, которые могут принимать большие значения теперь BigInteger
сделал реализацию апгрейта мощности автоматов

* теперь из json извлекается строка, а не int (теперь мы точно не выйдем за int)

* сделал форматирование денег и уменьшил начальную мощность клика (реально затягивает)

* пофиксил функцию форматирования денег

* поправил стиль везде кроме форматера

* изменил форматер. Функционал тот-же, но теперь понятнее

* теперь форматируются сколь угодно большие числа

* теперь форматируются сколь угодно большие числа

* вместо того, чтобы делать доп проверку в for в formatMoney просто укоротил его, это синонимично
зарефакторил слишком длинные строки

* зарефакторил automataThread и вынес рассчет суммарной мощности автоматов в класс магазина

* minor changes

+++ app/src/main/java/ru/sibur/android/garbagecollector/AutomataMoneyCalculator.java
@@ -1,9 +1,11 @@
 package ru.sibur.android.garbagecollector;
 
+import java.math.BigInteger;
+
 /**
  * Created by Олег on 23.03.2018.
  */
 
 public interface AutomataMoneyCalculator {
-    int calculateMoney (long timeDifference);
+    BigInteger calculateMoney (BigInteger timeDifference);
 }
